+++ e2e/modeling/test_dpr.py
@@ -1,7 +1,4 @@
-import os
 from typing import Tuple
-
-import logging
 from pathlib import Path
 
 import numpy as np
@@ -9,29 +6,17 @@ import pytest
 import torch
 from torch.utils.data import SequentialSampler
 from tqdm import tqdm
-from transformers import DPRQuestionEncoder, AutoTokenizer
+from transformers import AutoTokenizer
 
 from haystack.modeling.data_handler.dataloader import NamedDataLoader
 from haystack.modeling.data_handler.processor import TextSimilarityProcessor
 from haystack.modeling.model.biadaptive_model import BiAdaptiveModel
 from haystack.modeling.model.language_model import get_language_model, DPREncoder
 from haystack.modeling.model.prediction_head import TextSimilarityHead
-
-from haystack.nodes.retriever.dense import DensePassageRetriever
-
-from haystack.modeling.utils import set_all_seeds, initialize_device_settings
-from haystack.utils.early_stopping import EarlyStopping
-
-from ..conftest import SAMPLES_PATH
+from haystack.modeling.utils import initialize_device_settings
 
 
-def test_dpr_modules(caplog=None):
-    if caplog:
-        caplog.set_level(logging.CRITICAL)
-
-    set_all_seeds(seed=42)
-    devices, n_gpu = initialize_device_settings(use_cuda=True)
-
+def test_dpr_modules():
     # 1.Create question and passage tokenizers
     query_tokenizer = AutoTokenizer.from_pretrained(
         pretrained_model_name_or_path="facebook/dpr-question_encoder-single-nq-base", do_lower_case=True, use_fast=True
@@ -68,6 +53,7 @@ def test_dpr_modules(caplog=None):
 
     prediction_head = TextSimilarityHead(similarity_function="dot_product")
 
+    devices, _ = initialize_device_settings(use_cuda=True)
     model = BiAdaptiveModel(
         language_model1=question_language_model,
         language_model2=passage_language_model,
@@ -324,6 +310,13 @@ def test_dpr_processor(embed_title, passage_ids, passage_attns, use_fast, num_ha
                     "label": "hard_negative",
                     "external_id": "3643705",
                 },
+                # Empty title
+                {
+                    "title": "",
+                    "text": "Director Radio Iași); Dragoș-Liviu Vîlceanu; Mihnea-Adrian Vîlceanu; Nathalie-Teona",
+                    "label": "positive",
+                    "external_id": "b21eaeff-e08b-4548-b5e0-a280f6f4efef",
+                },
             ],
         },
         {
@@ -404,42 +397,6 @@ def test_dpr_processor(embed_title, passage_ids, passage_attns, use_fast, num_ha
             torch.eq(torch.tensor(feat[0]["label_ids"]), torch.tensor(labels[i])[: num_hard_negatives + 1])
         )
         assert len(torch.tensor(feat[0]["passage_segment_ids"]).nonzero()) == 0
-
-
-@pytest.mark.parametrize("use_fast", [False])
-@pytest.mark.parametrize("embed_title", [True, False])
-def test_dpr_processor_empty_title(use_fast, embed_title):
-    dict = {
-        "query": "what is a cat?",
-        "passages": [
-            {
-                "title": "",
-                "text": "Director Radio Iași); Dragoș-Liviu Vîlceanu; Mihnea-Adrian Vîlceanu; Nathalie-Teona",
-                "label": "positive",
-                "external_id": "b21eaeff-e08b-4548-b5e0-a280f6f4efef",
-            }
-        ],
-    }
-
-    query_tok = "facebook/dpr-question_encoder-single-nq-base"
-    query_tokenizer = AutoTokenizer.from_pretrained(query_tok, use_fast=use_fast)
-    passage_tok = "facebook/dpr-ctx_encoder-single-nq-base"
-    passage_tokenizer = AutoTokenizer.from_pretrained(passage_tok, use_fast=use_fast)
-    processor = TextSimilarityProcessor(
-        query_tokenizer=query_tokenizer,
-        passage_tokenizer=passage_tokenizer,
-        max_seq_len_query=256,
-        max_seq_len_passage=256,
-        data_dir="data/retriever",
-        train_filename="nq-train.json",
-        test_filename="nq-dev.json",
-        embed_title=embed_title,
-        num_hard_negatives=1,
-        label_list=["hard_negative", "positive"],
-        metric="text_similarity_metric",
-        shuffle_negatives=False,
-    )
-    _ = processor.dataset_from_dicts(dicts=[dict])
 
 
 def test_dpr_problematic():
@@ -975,170 +932,3 @@ def test_dpr_processor_save_load_non_bert_tokenizer(tmp_path: Path, query_and_pa
     # compare embeddings of model loaded from model hub and model loaded from disk that originated from a FARM style
     # model that was saved to disk earlier
     assert np.array_equal(all_embeddings["query"][0], all_embeddings3["query"][0])
-
-
-@pytest.mark.parametrize("document_store", ["memory"], indirect=True)
-def test_dpr_training(document_store, tmp_path):
-    retriever = DensePassageRetriever(
-        document_store=document_store,
-        query_embedding_model="facebook/dpr-question_encoder-single-nq-base",
-        passage_embedding_model="facebook/dpr-ctx_encoder-single-nq-base",
-        max_seq_len_query=8,
-        max_seq_len_passage=8,
-    )
-
-    save_dir = f"{tmp_path}/test_dpr_training"
-    retriever.train(
-        data_dir=str(SAMPLES_PATH / "dpr"),
-        train_filename="sample.json",
-        dev_filename="sample.json",
-        test_filename="sample.json",
-        n_epochs=1,
-        batch_size=1,
-        grad_acc_steps=1,
-        save_dir=save_dir,
-        evaluate_every=10,
-        embed_title=True,
-        num_positives=1,
-        num_hard_negatives=1,
-    )
-
-
-@pytest.mark.parametrize("document_store", ["memory"], indirect=True)
-def test_dpr_training_with_earlystopping(document_store, tmp_path):
-    retriever = DensePassageRetriever(
-        document_store=document_store,
-        query_embedding_model="facebook/dpr-question_encoder-single-nq-base",
-        passage_embedding_model="facebook/dpr-ctx_encoder-single-nq-base",
-        max_seq_len_query=8,
-        max_seq_len_passage=8,
-    )
-
-    save_dir = f"{tmp_path}/test_dpr_training"
-    retriever.train(
-        data_dir=str(SAMPLES_PATH / "dpr"),
-        train_filename="sample.json",
-        dev_filename="sample.json",
-        test_filename="sample.json",
-        n_epochs=1,
-        batch_size=1,
-        grad_acc_steps=1,
-        save_dir=save_dir,
-        evaluate_every=1,
-        embed_title=True,
-        num_positives=1,
-        num_hard_negatives=1,
-        early_stopping=EarlyStopping(save_dir=save_dir),
-    )
-
-
-# TODO fix CI errors (test pass locally or on AWS, next steps: isolate PyTorch versions once FARM dependency is removed)
-# def test_dpr_training():
-#     batch_size = 1
-#     n_epochs = 1
-#     distributed = False  # enable for multi GPU training via DDP
-#     evaluate_every = 1
-#     question_lang_model = "microsoft/MiniLM-L12-H384-uncased"
-#     passage_lang_model = "microsoft/MiniLM-L12-H384-uncased"
-#     do_lower_case = True
-#     use_fast = True
-#     similarity_function = "dot_product"
-#
-#     device, n_gpu = initialize_device_settings(use_cuda=False)
-#
-#     query_tokenizer = get_tokenizer(pretrained_model_name_or_path=question_lang_model,
-#                                      do_lower_case=do_lower_case, use_fast=use_fast)
-#     passage_tokenizer = get_tokenizer(pretrained_model_name_or_path=passage_lang_model,
-#                                        do_lower_case=do_lower_case, use_fast=use_fast)
-#     label_list = ["hard_negative", "positive"]
-#
-#     processor = TextSimilarityProcessor(query_tokenizer=query_tokenizer,
-#                                         passage_tokenizer=passage_tokenizer,
-#                                         max_seq_len_query=10,
-#                                         max_seq_len_passage=10,
-#                                         label_list=label_list,
-#                                         metric="text_similarity_metric",
-#                                         data_dir="samples/dpr/",
-#                                         train_filename="sample.json",
-#                                         dev_filename="sample.json",
-#                                         test_filename=None,
-#                                         embed_title=True,
-#                                         num_hard_negatives=1,
-#                                         dev_split=0,
-#                                         max_samples=2)
-#
-#     data_silo = DataSilo(processor=processor, batch_size=batch_size, distributed=False)
-#
-#     question_language_model = get_language_model(pretrained_model_name_or_path=question_lang_model,
-#                                                  language_model_class="DPRQuestionEncoder")
-#     passage_language_model = get_language_model(pretrained_model_name_or_path=passage_lang_model,
-#                                                 language_model_class="DPRContextEncoder")
-#
-#     prediction_head = TextSimilarityHead(similarity_function=similarity_function)
-#
-#     model = BiAdaptiveModel(
-#         language_model1=question_language_model,
-#         language_model2=passage_language_model,
-#         prediction_heads=[prediction_head],
-#         embeds_dropout_prob=0.1,
-#         lm1_output_types=["per_sequence"],
-#         lm2_output_types=["per_sequence"],
-#         device=device,
-#     )
-#
-#     model, optimizer, lr_schedule = initialize_optimizer(
-#         model=model,
-#         learning_rate=1e-5,
-#         optimizer_opts={"name": "TransformersAdamW", "correct_bias": True, "weight_decay": 0.0, \
-#                         "eps": 1e-08},
-#         schedule_opts={"name": "LinearWarmup", "num_warmup_steps": 100},
-#         n_batches=len(data_silo.loaders["train"]),
-#         n_epochs=n_epochs,
-#         grad_acc_steps=1,
-#         device=device,
-#         distributed=distributed
-#     )
-#
-#     trainer = Trainer(
-#         model=model,
-#         optimizer=optimizer,
-#         data_silo=data_silo,
-#         epochs=n_epochs,
-#         n_gpu=n_gpu,
-#         lr_schedule=lr_schedule,
-#         evaluate_every=evaluate_every,
-#         device=device,
-#     )
-#
-#     trainer.train()
-#
-#     ######## save and load model again
-#     save_dir = Path("testsave/dpr-model")
-#     model.save(save_dir)
-#     del model
-#
-#     model2 = BiAdaptiveModel.load(save_dir, device=device)
-#     model2, optimizer2, lr_schedule = initialize_optimizer(
-#         model=model2,
-#         learning_rate=1e-5,
-#         optimizer_opts={"name": "TransformersAdamW", "correct_bias": True, "weight_decay": 0.0, \
-#                         "eps": 1e-08},
-#         schedule_opts={"name": "LinearWarmup", "num_warmup_steps": 100},
-#         n_batches=len(data_silo.loaders["train"]),
-#         n_epochs=n_epochs,
-#         grad_acc_steps=1,
-#         device=device,
-#         distributed=distributed
-#     )
-#     trainer2 = Trainer(
-#         model=model2,
-#         optimizer=optimizer,
-#         data_silo=data_silo,
-#         epochs=n_epochs,
-#         n_gpu=n_gpu,
-#         lr_schedule=lr_schedule,
-#         evaluate_every=evaluate_every,
-#         device=device,
-#     )
-#
-#     trainer2.train()
